Added abstain voting icon.  Reorganized types into Shared folder.
