+++ src/components/Shared/AbstainIcon.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import "@councildataproject/cdp-design/dist/colors.css";
+
+const AbstainIcon = () => {
+  return (
+    <svg
+      className="cdp-light-purple"
+      fill="none"
+      stroke="currentColor"
+      viewBox="0 0 24 24"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width={2}
+        d="M5 12h.01M12 12h.01M19 12h.01M6 12a1 1 0 11-2 0 1 1 0 012 0zm7 0a1 1 0 11-2 0 1 1 0 012 0zm7 0a1 1 0 11-2 0 1 1 0 012 0z"
+      />
+    </svg>
+  );
+};
+
+export default AbstainIcon;
