+++ src/components/Shared/DecisionResult.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import LegislationAdoptedIcon from "../Cards/LegislationCard/LegislationAdoptedIcon";
 import LegislationRejectedIcon from "../Cards/LegislationCard/LegislationRejectedIcon";
 import LegislationInProgressIcon from "../Cards/LegislationCard/LegislationInProgressIcon";
+import AbstainIcon from "../Shared/AbstainIcon";
 import { MATTER_STATUS_DECISION } from "../../constants/ProjectConstants";
 import { VOTE_DECISION } from "../../constants/ProjectConstants";
 import "@mozilla-protocol/core/protocol/css/protocol.css";
@@ -19,6 +20,9 @@ const DecisionResult = ({ result }: DecisionResultProps) => {
       break;
     case VOTE_DECISION.REJECT:
       statusIcon = <LegislationRejectedIcon />;
+      break;
+    case VOTE_DECISION.ABSTAIN:
+      statusIcon = <AbstainIcon />;
       break;
     case MATTER_STATUS_DECISION.REJECTED:
       statusIcon = <LegislationRejectedIcon />;
