+++ src/components/Shared/Types/Matter.tsx
@@ -1,0 +1,6 @@
+export type Matter = {
+  /** the name of the matter, usually a legislation name */
+  name: string;
+  /** the id of the matter */
+  id: string;
+};
